% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epiwidgets-package.R
\docType{package}
\name{epiwidgets-package}
\alias{epiwidgets-package}
\title{epiwidgets: htmlwidgets using D3 and three.js}
\description{
To learn more about epiwidget, start with the vignettes:
\code{browseVignettes(package = "epiwidgets")}
}

