% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixcube.R
\name{matrixcube}
\alias{matrixcube}
\title{matrixcube Three.js matrixcube widget.}
\usage{
matrixcube(x, y, z, height = NULL, width = NULL, axis = TRUE,
  axis.labels = NULL, tick.labels = NULL, color = "steelblue",
  stroke = "black", size = 1, flip.time = FALSE, grid = TRUE,
  renderer = c("auto", "canvas", "webgl", "webgl-buffered"), pch,
  browser = FALSE)
}
\description{
Based on the 3D scatterplot widget from rthreejs.
}

