% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/odewidget.R
\name{odewidgetOutput}
\alias{odewidgetOutput}
\title{Widget output function for use in Shiny}
\usage{
odewidgetOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

