% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/packwidget.R
\name{packwidgetOutput}
\alias{packwidgetOutput}
\title{Widget output function for use in Shiny}
\usage{
packwidgetOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

