% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/odewidget.R
\name{renderOdewidget}
\alias{renderOdewidget}
\title{Widget render function for use in Shiny}
\usage{
renderOdewidget(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

