% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/packwidget.R
\name{renderPackwidget}
\alias{renderPackwidget}
\title{Widget render function for use in Shiny}
\usage{
renderPackwidget(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

