% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/treewidget.R
\name{renderTreewidget}
\alias{renderTreewidget}
\title{renderTreewidget Helper function for treeshiny}
\usage{
renderTreewidget(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{The function to be rendered}

\item{env}{The environment for rendering.}

\item{quoted}{Should the output be quoted?}
}
\description{
renderTreewidget Helper function for treeshiny
}
\seealso{
treewidgetOutput
}

