% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/treewidget.R
\name{treeshiny}
\alias{treeshiny}
\title{treeshiny A treewidget as a Shiny app}
\usage{
treeshiny(nwk)
}
\arguments{
\item{nwk}{Either a Newick string or a \code{phylo} or \code{multiPhylo} object.}
}
\description{
treeshiny A treewidget as a Shiny app
}
\examples{
\dontrun{
library(ape)
data(bird.orders)
treeshiny(bird.orders)
}
}
\seealso{
treewidget
}

