% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/treewidget.R
\name{treewidgetOutput}
\alias{treewidgetOutput}
\title{treewidgetOutput Helper function for treeshiny}
\usage{
treewidgetOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{The output identifier.}

\item{width}{The width of the device.}

\item{height}{The height of the device.}
}
\description{
treewidgetOutput Helper function for treeshiny
}
\seealso{
renderTreewidget
}

